some script

+++ cluster-import/performance/cleanup.sh
@@ -1,3 +1,8 @@
 #!/bin/bash
 
-kubectl get managedclusters | grep test | awk '{print $1}' | xargs kubectl delete managedclusters
+if [ "$1"x = "force"x ]; then
+    kubectl get managedclusters | grep -v NAME | awk '{print $1}' | xargs kubectl patch managedclusters -p '{"metadata":{"finalizers": []}}' --type=merge
+    exit 0
+fi
+
+kubectl get managedclusters | grep test | awk '{print $1}' | xargs kubectl delete managedclusters
